requirements.txt
openai>=1.0
tenacity>=8.0
httpx>=0.25
pytest-asyncio>=0.23
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
opensearch-py>=2.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
langgraph>=0.0.10
